allow customized invoker

+++ transport/report.go
@@ -3,26 +3,26 @@ package transport
 type Report struct {
 	I string
 	N string
-	S int
+	S int16
 	E *Error
 	R []interface{}
 }
 
 var Status = struct {
-	None     int
-	Sent     int
-	Progress int
-	Done     int
-	Fail     int
-	Count    int // this field should always the last one
+	None     int16
+	Sent     int16
+	Progress int16
+	Done     int16
+	Fail     int16
+	Count    int16 // this field should always the last one
 
 	// these fields are for test
-	Test1 int
-	Test2 int
-	Test3 int
-	Test4 int
-	Test5 int
-	Test6 int
+	Test1 int16
+	Test2 int16
+	Test3 int16
+	Test4 int16
+	Test5 int16
+	Test6 int16
 }{
 	0, 1, 2, 3, 4, 5,
 
@@ -35,7 +35,7 @@ var Status = struct {
 //
 func (r *Report) ID() string            { return r.I }
 func (r *Report) Name() string          { return r.N }
-func (r *Report) Status() int           { return r.S }
+func (r *Report) Status() int16         { return r.S }
 func (r *Report) Err() *Error           { return r.E }
 func (r *Report) Return() []interface{} { return r.R }
