+++ transport/report_test.go
@@ -22,7 +22,7 @@ func TestReportMarshal(t *testing.T) {
 	err = json.Unmarshal(body, &r)
 	ass.Nil(err)
 	if err == nil {
-		ass.Equal(101, r.Status())
+		ass.Equal(int16(101), r.Status())
 		ass.Equal("test_id", r.ID())
 		ass.Equal(int64(102), r.Err().Code())
 		ass.Equal("test error", r.Err().Msg())
